Agregando validaciones de asignacion y de curso

+++ src/controllers/curso.controllers.js
@@ -4,19 +4,24 @@ function agregarCurso (req, res){
     var parametros = req.body;
     var modeloCurso = new curso();
 
-    if(parametros.nombre){
-        modeloCurso.nombre = parametros.nombre;
-        modeloCurso.idMaestro = req.user.sub;
+    if(req.user.rol == "ROL_ALUMNO"){
+        return res.status(500).send({mensaje: "No tiene los permisos para asignar un curso"});
+    } else {
 
-        modeloCurso.save((err, cursoGuardado) => {
-            if(err) return res.status(500).send({ mensaje : 'Error en la peticion' });
-            if(!cursoGuardado) return res.status(500).send({ mensaje: 'Error al agregar el curso' });
+            if(parametros.nombre){
+                modeloCurso.nombre = parametros.nombre;
+                modeloCurso.idMaestro = req.user.sub;
 
-            return res.status(200).send({ curso: cursoGuardado})
-        })
+                modeloCurso.save((err, cursoGuardado) => {
+                    if(err) return res.status(500).send({ mensaje : 'Error en la peticion' });
+                    if(!cursoGuardado) return res.status(500).send({ mensaje: 'Error al agregar el curso' });
 
-    }else{
-        return res.status(500).send({message: "Debe ingresar los parametros obligatorios"});
+                    return res.status(200).send({ curso: cursoGuardado})
+                })
+
+            }else{
+                return res.status(500).send({message: "Debe ingresar los parametros obligatorios"});
+            }
     }
 }
 
@@ -25,27 +30,42 @@ function editarCurso (req, res) {
     var parametros = req.body;
     var idCurso = req.params.idCurso;
 
-    curso.findByIdAndUpdate(idCurso, parametros, {new : true}, (err, cursoEditado) => {
 
-        if(err)return res.status(500).send({mensaje: 'Error en la petición'});
-        if(!cursoEditado)return res.status(404).send({mensaje: 'Error al editar el curso'});
+    if(req.user.rol == "ROL_ALUMNO"){
+        return res.status(500).send({mensaje: "No tiene los permisos para asignar un curso"});
+    } else {
 
-        return res.status(200).send({curso: cursoEditado});
-    })
+        curso.findByIdAndUpdate(idCurso, parametros, {new : true}, (err, cursoEditado) => {
 
+            if(err)return res.status(500).send({mensaje: 'Error en la petición'});
+            if(!cursoEditado)return res.status(404).send({mensaje: 'Error al editar el curso'});
+    
+            return res.status(200).send({curso: cursoEditado});
+        })
+    
+        
+    }
 
 }
 
 function eliminarCurso (req, res) {
     var idCurso = req.params.idCurso;
 
-    curso.findByIdAndDelete(idCurso, (err, cursoEliminado) => {
 
-        if(err) return res.status(500).send({mensaje: 'Error en la petición'});
-        if(!cursoEliminado)return res.status(404).send({mensaje: 'Error al eliminar el curso'});
 
-        return res.status(200).send({curso: cursoEliminado});
-    })
+    if(req.user.rol == "ROL_ALUMNO"){
+        return res.status(500).send({mensaje: "No tiene los permisos para asignar un curso"});
+    } else {
+
+        curso.findByIdAndDelete(idCurso, (err, cursoEliminado) => {
+
+            if(err) return res.status(500).send({mensaje: 'Error en la petición'});
+            if(!cursoEliminado)return res.status(404).send({mensaje: 'Error al eliminar el curso'});
+    
+            return res.status(200).send({curso: cursoEliminado});
+        })
+        
+    }
 }
 
 function obtenerCurso (req, res){
